+++ lib/cocoapods-downloader/git.rb
@@ -22,6 +22,23 @@ module Pod
         end
       end
 
+      def self.preprocess_options(options)
+        return options unless options[:branch]
+
+        command = ['ls-remote',
+                   options[:git],
+                   options[:branch]]
+        output = Git.execute_command('git', command)
+        match = /^([a-z0-9]*)\t.*/.match(output)
+
+        return options if match.nil?
+
+        options[:commit] = match[1]
+        options.delete(:branch)
+
+        options
+      end
+
       private
 
       # @!group Base class hooks