+++ spec/git_spec.rb
@@ -203,6 +203,26 @@ module Pod
           should.not.raise { downloader.download }
         end
       end
+
+      describe '::preprocess_options' do
+        it 'skips non-branch requests' do
+          options = { :git => fixture_url('git-repo'), :commit => 'aaaa' }
+          new_options = Downloader.preprocess_options(options)
+          options.should == new_options
+        end
+
+        it 'resolves master to a commit' do
+          options = { :git => fixture_url('git-repo'), :branch => 'master' }
+          new_options = Downloader.preprocess_options(options)
+          new_options[:commit].should == '98cbf14201a78b56c6b7290f6cac840a7597a1c2'
+        end
+
+        it 'ignores invalid branches' do
+          options = { :git => fixture_url('git-repo'), :branch => 'aaaa' }
+          new_options = Downloader.preprocess_options(options)
+          new_options[:branch].should == 'aaaa'
+        end
+      end
     end
   end
 end